+++ src/onegov.activity/HISTORY.rst
@@ -1,5 +1,9 @@
 Changelog
 ---------
+
+- Adds a subscription token to the attendee.
+  [href]
+
 1.6.1 (2018-02-12)
 ~~~~~~~~~~~~~~~~~~~